fixing unsigned integer problems, refs #1316
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16504 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicleContainer.cpp
@@ -122,7 +122,7 @@ void
 MSVehicleContainer::addReplacing(const VehicleDepartureVector& x) {
     if (isFull()) {
         std::vector<VehicleDepartureVector> array2((array.size() - 1) * 2 + 1, VehicleDepartureVector());
-        for (size_t i = array.size(); i-- > 0;) {
+        for (int i = (int)array.size(); i-- > 0;) {
             assert(array2.size() > i);
             array2[i] = array[i];
         }
